Improve perspective transformation code in soft shadow volumes article
I figured this one out before writing the blog, but the change somehow
got lost in the git history (and my brain). But yes, the `abs()`
solution was certainly not correct. With this old solution if a triangle
cut through the ground, the shadow right next to it was already quite
blurred.